Removed all warnings.

+++ src/1.cc
@@ -33,7 +33,7 @@ void hexDecode(const char* input, uint8_t* out) {
 
   std::string myString(input);
 
-  for (int i = 0; i < (myString.size()) / 2; i++) {
+  for (int i = 0; i < ((int)myString.size()) / 2; i++) {
     std::string subStr = myString.substr(i * 2, 2);
     out[i] = strtoul(subStr.c_str(), 0, 16);
   }
